Convert addresses correctly.
