+++ apps/explorer/test/explorer/celo/events/validator_group_vote_activated_event_test.exs
@@ -125,8 +125,8 @@ defmodule Explorer.Celo.Events.ValidatorGroupVoteActivatedEventTest do
 
       assert result.value == 66980000000000000000
       assert result.units == 6136281451163456507329304650157103347504
-      assert result.account == "0x88c1c759600ec3110af043c183a2472ab32d099c"
-      assert result.group == "0x47b2db6af05a55d42ed0f3731735f9479abf0673"
+      assert to_string(result.account) == "0x88c1c759600ec3110af043c183a2472ab32d099c"
+      assert to_string(result.group) == "0x47b2db6af05a55d42ed0f3731735f9479abf0673"
     end
   end
 end
